+++ lib/dubp/user-docs/src/documents/membership/v10.rs
@@ -194,7 +194,7 @@ impl MembershipDocumentV10 {
 impl Document for MembershipDocumentV10 {
     type PublicKey = PubKey;
 
-    fn version(&self) -> u16 {
+    fn version(&self) -> usize {
         10
     }
